soc: aspeed: mctp: Remove the check of bdf value
The original condition where bdf is zero is treated as an error because
our MCTP is only over the PCIe EP port. However, if the MCTP is over the
PCIe RC, the bdf being zero is normal. In fact, the driver doesn't need to
concern itself with the bdf; checking the PCIe Link status is sufficient.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I66ee42adfe5fce5e4aed8436847ccae8f846e856