i3c: mipi-i3c-hci: free the data when xfer complete
The transfer in target mode is initiated by the i3c primary, so the data
for the transfer shouldn't be freed until the target receives the
interrupt indicating that the data has been used.
Additionally, this patch removes support for the priv_xfer with target
mode due to the aforementioned situation.

Signed-off-by: Billy Tsai <billy_tsai@aspeedtech.com>
Change-Id: I3380809933d60b37791e466f5a137a3e5b6707c5