dtsi: aspeed_g7/-a1: Fix the typo.
In dts, values for status property are
1. "okay" Indicates the device is operational.

2. "disabled" Indicates that the device is not presently operational, but
it might become operational in the future (for example, something is not
plugged in, or switched off). Refer to the device binding for details on
what disabled means for a given device.

3. "reserved" Indicates that the device is operational, but should not be
used. Typically this is used for devices that are controlled by another
software component, such as platform firmware.

4. "fail" Indicates that the device is not operational. A serious error
was detected in the device, and it is unlikely to become operational
without repair.

5. "fail-sss" Indicates that the device is not operational. A serious
error was detected in the device and it is unlikely to become operational
without repair. The sss portion of the value is specific to the device and
indicates the error condition detected.

Change-Id: I53d814da21f758bfc53a7937184156783883ef1f
Signed-off-by: Kevin Chen <kevin_chen@aspeedtech.com>